feature/SCAT 2302 - file Upload Validations and error handling
