+++ src/main/features/rfi/controller/upload.ts
@@ -9,7 +9,8 @@ import FormData from 'form-data'
 import fileSystem from 'fs'
 import util from 'util'
 import stream from 'stream';
-
+import {FileValidations} from '../util/file/filevalidations'
+import {FILEUPLOADHELPER} from '../helpers/upload'
 
 // RFI Upload document
 /**
@@ -20,97 +21,7 @@ import stream from 'stream';
  */
 
 export const GET_UPLOAD_DOC: express.Handler = async (req: express.Request, res: express.Response) => {
-      const lotId = req.session?.lotId;
-      const {SESSION_ID} = req.cookies;
-      const agreementLotName = req.session.agreementLotName;
-      const ProjectId = req.session['projectId'];
-      const EventId = req.session['eventId'];
-      const {file_id} =req.query;
-      if(file_id !== undefined){
-            try {
-            const FileDownloadURL = `/tenders/projects/${ProjectId}/events/${EventId}/documents/${file_id}`
-            const FetchDocuments = await DynamicFrameworkInstance.file_dowload_Instance(SESSION_ID).get(FileDownloadURL, {responseType: 'blob'});
-            const file = FetchDocuments;
-            const fileName = file.headers["content-disposition"].split("filename=")[1].split('"').join("");
-            const type = file.headers['content-type']
-            const pipeline = util.promisify(stream.pipeline);
-            const charset = type.split(";")[1].split("charset=");
-            const fileData = file.data;
-            const fileTemporaryPath = 'uploads/downloads/'+fileName;
-            await pipeline(fileData, fileSystem.createWriteStream(fileTemporaryPath));
-
-            res.download(fileTemporaryPath, (err)=>{
-              if(err){
-                delete error?.config?.['headers'];
-                const Logmessage = {
-                  Person_id: TokenDecoder.decoder(SESSION_ID),
-                  error_location: `${req.headers.host}${req.originalUrl}`,
-                  sessionId: 'null',
-                  error_reason: 'Error -Downloading File to the local Storage',
-                  exception: error,
-                };
-                const Log = new LogMessageFormatter(
-                  Logmessage.Person_id,
-                  Logmessage.error_location,
-                  Logmessage.sessionId,
-                  Logmessage.error_reason,
-                  Logmessage.exception,
-                );
-                fileSystem.unlinkSync(fileTemporaryPath);
-                LoggTracer.errorTracer(Log, res);
-              }else{
-                fileSystem.unlinkSync(fileTemporaryPath);
-              }
-          
-          })
-
-        } catch (error) {
-          delete error?.config?.['headers'];
-          const Logmessage = {
-            Person_id: TokenDecoder.decoder(SESSION_ID),
-            error_location: `${req.headers.host}${req.originalUrl}`,
-            sessionId: 'null',
-            error_reason: 'File uploading Causes Problem in RFI  - Tenders Api throws error',
-            exception: error,
-          };
-          const Log = new LogMessageFormatter(
-            Logmessage.Person_id,
-            Logmessage.error_location,
-            Logmessage.sessionId,
-            Logmessage.error_reason,
-            Logmessage.exception,
-          );
-          LoggTracer.errorTracer(Log, res);
-        }
-      }
-      else{
-        try {
-          const FileuploadBaseUrl = `/tenders/projects/${ProjectId}/events/${EventId}/documents`
-          const FetchDocuments = await DynamicFrameworkInstance.Instance(SESSION_ID).get(FileuploadBaseUrl);
-          const FETCH_FILEDATA = FetchDocuments.data;
-          const windowAppendData = { lotId, agreementLotName, data: cmsData, files: FETCH_FILEDATA }
-          res.render('uploadDocument', windowAppendData);
-    } catch (error) {
-          delete error?.config?.['headers'];
-          const Logmessage = {
-            Person_id: TokenDecoder.decoder(SESSION_ID),
-            error_location: `${req.headers.host}${req.originalUrl}`,
-            sessionId: 'null',
-            error_reason: 'File uploading Causes Problem in RFI  - Tenders Api throws error',
-            exception: error,
-          };
-          const Log = new LogMessageFormatter(
-            Logmessage.Person_id,
-            Logmessage.error_location,
-            Logmessage.sessionId,
-            Logmessage.error_reason,
-            Logmessage.exception,
-          );
-          LoggTracer.errorTracer(Log, res);
-        }
-      }
-
-     
+  FILEUPLOADHELPER(req, res, false, null)
 }
 
 /**
@@ -132,6 +43,7 @@ export const GET_UPLOAD_DOC: express.Handler = async (req: express.Request, res:
   const EventId = req.session['eventId'];
   const FILE_PUBLISHER_BASEURL = `/tenders/projects/${ProjectId}/events/${EventId}/documents`
 
+  const FileFilterArray = [];
 
   if (rfi_file_started) {
       const {
@@ -142,10 +54,20 @@ export const GET_UPLOAD_DOC: express.Handler = async (req: express.Request, res:
       if (multipleFileCheck) {
           for (const file of rfi_offline_document) {
               const fileName = file.name;
-              const fileTemporaryPath = 'uploads/upload/'
-              file.mv(fileTemporaryPath + fileName, async (err) => {
+              const fileTemporaryPath = 'uploads/upload/';
+
+              const fileMimeType = file.mimetype;
+              const fileSize = file.size;
+    
+              const validateMimeType = FileValidations.formatValidation(fileMimeType);
+              const validateFileSize = FileValidations.sizeValidation(fileSize);
+              
+              if(validateMimeType && validateFileSize)
+              {
+                file.mv(fileTemporaryPath + fileName, async (err) => {
                   if (err) {
-                      res.render('error/500')
+                    LoggTracer.errorLogger(res, error, `${req.headers.host}${req.originalUrl}`, null,
+                    TokenDecoder.decoder(SESSION_ID), "Multiple File Upload Error", false)
                   }
                   const formData = new FormData();
                   const formDataReadPath = fileTemporaryPath + fileName;
@@ -167,14 +89,37 @@ export const GET_UPLOAD_DOC: express.Handler = async (req: express.Request, res:
                           TokenDecoder.decoder(SESSION_ID), "Multiple File Upload Error", false)
                   }
               });
+               
+
+              }
+              else{
+                FileFilterArray.push({
+                  "href": "#documents_upload",
+                  "text" : fileName
+                })
+              }           
 
           }
 
-          res.redirect('/rfi/upload-doc')
+          if(FileFilterArray.length > 0){
+            FILEUPLOADHELPER(req, res, true, FileFilterArray);
+          }
+          else  res.redirect('/rfi/upload-doc')
+
+         
       } else {
           const fileName = rfi_offline_document.name;
-          const fileTemporaryPath = 'uploads/'
-          rfi_offline_document.mv(fileTemporaryPath + fileName, async (err) => {
+          const fileTemporaryPath = 'uploads/upload/'
+
+          const fileMimeType = rfi_offline_document.mimetype;
+          const fileSize = rfi_offline_document.size;
+
+          const validateMimeType = FileValidations.formatValidation(fileMimeType);
+          const validateFileSize = FileValidations.sizeValidation(fileSize);
+
+          if(validateMimeType && validateFileSize)
+          {
+            rfi_offline_document.mv(fileTemporaryPath + fileName, async (err) => {
               if (err) {
                   res.render('error/500')
               }
@@ -212,6 +157,18 @@ export const GET_UPLOAD_DOC: express.Handler = async (req: express.Request, res:
                   LoggTracer.errorTracer(Log, res);
               }
           });
+          }
+
+          else{
+            FileFilterArray.push({
+              "href": "#documents_upload",
+              "text" : fileName
+            })
+
+            FILEUPLOADHELPER(req, res, true, FileFilterArray);
+
+          }
+         
       }
 
   } else res.render('error/500')
