+++ src/main/features/rfi/util/file/filevalidations.ts
@@ -1,0 +1,17 @@
+//@ts-nocheck
+import {FileMimeType} from './mimetype'
+
+export class FileValidations{
+
+        static sizeValidation = (fileSize: number) => {
+        const KbsInGbs = 1000000;
+        if(fileSize < KbsInGbs) return true;
+        else return false;
+        }
+
+        static formatValidation = (mimeType ) => {
+                const allMimeTypes = Object.values(FileMimeType);
+                const checkForFileMimeType = allMimeTypes.filter(aMimeType => aMimeType === mimeType).length > 0;
+               return checkForFileMimeType;
+        }
+}
