+++ src/main/features/rfi/util/file/mimetype.ts
@@ -4,14 +4,12 @@ export const FileMimeType = {
     "doc": "application/msword",
     "docx": "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
     "jpg": "image/jpeg",
-    "kml": "",
     "ods": "application/vnd.oasis.opendocument.spreadsheet",
     "odt": "application/vnd.oasis.opendocument.text",
     "pdf": "application/pdf",
     "png": "image/png",
     "ppt": "application/vnd.ms-powerpoint",
     "pptx": "application/vnd.openxmlformats-officedocument.presentationml.presentation",
-    "rdf": "", 
     "rtf": "application/rtf",
     "txt": "text/plain",
     "xls": "application/vnd.ms-excel",
